Swap tests from SDK to oteljava
